+++ packages/lumx-core/src/js/components/Combobox/constants.ts
@@ -1,0 +1,3 @@
+import { VHSize } from '../../utils/browser/css/types';
+
+export const DEFAULT_COMBOBOX_POPOVER_MAX_HEIGHT: VHSize = '80vh';